Move Provider To App
